refactor: adequa o nome do pacote ao padrão de nomenclatura
