+++ api-backend/src/main/java/br/com/retermais/ReterMaisApplication.java
@@ -1,4 +1,4 @@
-package br.com.ReterMais;
+package br.com.retermais;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
