+++ api-backend/src/main/java/br/com/retermais/dtos/RequestPythonDTO.java
@@ -1,11 +1,10 @@
-package br.com.ReterMais.dtos;
+package br.com.retermais.dtos;
 
-import br.com.ReterMais.model.*;
+import br.com.retermais.model.*;
 
 import java.math.BigDecimal;
 
-public record RequestDTO(
-        String idCliente,
+public record RequestPythonDTO(
         Genero genero,
         Integer idoso,
         Integer parceiro,
