+++ api-backend/src/main/java/br/com/retermais/model/ProtecaoDispositivo.java
@@ -1,15 +1,15 @@
-package br.com.ReterMais.model;
+package br.com.retermais.model;
 
 import com.fasterxml.jackson.annotation.JsonValue;
 
-public enum SuporteTecnico {
+public enum ProtecaoDispositivo {
     NAO("nao"),
     SEM_INTERNET("sem_internet"),
     SIM("sim");
 
     private final String descricao;
 
-    SuporteTecnico(String descricao) {
+    ProtecaoDispositivo(String descricao) {
         this.descricao = descricao;
     }
