+++ api-backend/src/main/java/br/com/retermais/service/ReterMaisService.java
@@ -1,8 +1,8 @@
-package br.com.ReterMais.service;
+package br.com.retermais.service;
 
 
-import br.com.ReterMais.dtos.RequestDTO;
-import br.com.ReterMais.dtos.ResponsePythonDTO;
+import br.com.retermais.dtos.RequestDTO;
+import br.com.retermais.dtos.ResponsePythonDTO;
 import org.springframework.stereotype.Service;
 
 @Service
