+++ api-backend/src/main/resources/application.properties
@@ -1,1 +1,1 @@
-spring.application.name=ReterMais
+spring.application.name=retermais
