+++ api-backend/src/test/java/br/com/retermais/ReterMaisApplicationTests.java
@@ -1,4 +1,4 @@
-package br.com.ReterMais;
+package br.com.retermais;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
@@ -6,8 +6,8 @@ import org.springframework.boot.test.context.SpringBootTest;
 @SpringBootTest
 class ReterMaisApplicationTests {
 
-	@Test
-	void contextLoads() {
-	}
+    @Test
+    void contextLoads() {
+    }
 
 }
